Split session interaction IPC handlers
